🚪 users can now sign out, sidebar links now react to he current bg, merged in tyes old side bar code.
