+++ react-app/src/Components/Auth/index.js
@@ -1,3 +1,4 @@
 export { default as AuthLoader } from './Auth'
 export { default as Login } from './Login'
 export { ProtectedRoute } from './ProtectedRoute'
+export { signout } from './signout'
