+++ react-app/src/Components/sideBar/index.js
@@ -1,0 +1,1 @@
+export { default as SideBar } from './Sidebar'
