+++ react-app/src/utils/auth.js
@@ -44,13 +44,6 @@ export default class Auth {
 		return new Date().getTime() < expiresAt;
 	}
 
-	logout() {
-		localStorage.removeItem('access_token');
-		localStorage.removeItem('id_token');
-		localStorage.removeItem('expires_at');
-		window.location.pathname = '/';
-	}
-
 	getProfile() {
 		if (localStorage.getItem('id_token')) {
 			return jwtDecode(localStorage.getItem('id_token'));
